+++ editor/terrain/voxel_terrain_editor_plugin.h
@@ -26,6 +26,7 @@ private:
 
 	enum MenuID {
 		MENU_RESTART_STREAM,
+		MENU_REMESH,
 		MENU_ABOUT
 	};